% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsolvetk.R
\name{ev_seq}
\alias{ev_seq}
\title{Schedule a sequence of event objects.}
\usage{
ev_seq(..., .dots = list())
}
\arguments{
\item{...}{event objects}

\item{.dots}{list of event objects}
}
\description{
Schedule a sequence of event objects.
}
