% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgoptim.R
\name{mrgoptim}
\alias{mrgoptim}
\title{Fit a model to data using maximum likelihood objective function
with newuoa or optim}
\usage{
mrgoptim(mod, output = "ipred", var = "var", prms, v_prms, cov = TRUE,
  restarts = 0, method = c("newuoa", "Nelder-Mead", "BFGS", "CG",
  "L-BFGS-B", "SANN", "Brent"), options = list(maxit = 10000))
}
\arguments{
\item{mod}{a model object}

\item{output}{a character string containing name of prediction output column}

\item{var}{a character vector containing name of variance output column of preditcions}

\item{prms}{a character vector of parameters to fit}

\item{v_prms}{a character vector of variance parameters to fit}

\item{cov}{logical, perform covariance step}

\item{restarts}{number of times to restart fit with final estimates of previous optimization}

\item{method}{method to be used: "newuoa","Nelder-Mead", "BFGS", "CG", "L-BFGS-B", "SANN", "Brent"}

\item{options}{options passed to optim function, see ?optim for more information}
}
\description{
Fit a model to data using maximum likelihood objective function
with newuoa or optim
}
