% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsolvetk.R
\name{sens_unif}
\alias{sens_unif}
\alias{sens_norm}
\alias{sens_seq}
\alias{sens_grid}
\alias{sens_covset}
\alias{sens_unif_}
\alias{sens_norm_}
\alias{sens_seq_}
\alias{sens_unif_idata}
\alias{sens_norm_idata}
\title{Perform sensititity analysis on model parameters.}
\usage{
sens_unif(mod, n = 100, pars = names(param(mod)), ...)

sens_norm(mod, n = 100, pars = names(param(mod)), ...)

sens_seq(mod, n = 100, ...)

sens_grid(mod, n = 100, ...)

sens_covset(mod, covset, n = 100, ...)

sens_unif_(x, .dots)

sens_norm_(mod, .dots)

sens_seq_(mod, .dots)

sens_unif_idata(pars, lower = 0.2, upper = 3, n = 100, spread = TRUE,
  ...)

sens_norm_idata(pars, cv, n = 100, log = TRUE, spread = TRUE, ...)
}
\arguments{
\item{mod}{the model object}

\item{n}{the number of replicates to simulate}

\item{pars}{character vector or comma-separated string of 
model parameters to simulate}
}
\description{
Perform sensititity analysis on model parameters.
}
