% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsolvetk.R
\name{sens_covset}
\alias{sens_covset}
\title{Sensitivity analysis using covset objects.}
\usage{
sens_covset(mod, covset, n = 100, ...)
}
\arguments{
\item{mod}{the model object}

\item{covset}{a covset object}

\item{n}{the number of replicates to simulate}

\item{...}{passed to mutate_random and mrgsim}
}
\description{
Sensitivity analysis using covset objects.
}
\seealso{
\code{\link{sens_unif}} \code{\link{sens_norm}}
}
