% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens_each.R
\name{sens_each}
\alias{sens_each}
\title{Sensitivity analysis one-by-one}
\usage{
sens_each(mod, ...)
}
\arguments{
\item{mod}{a model object}

\item{...}{numeric vectors with names matching model parameter names}
}
\description{
Sensitivity analysis one-by-one
}
\examples{
mod <- mrgsolve:::house() \%>\% ev(amt = 100)

mod \%>\%
  sens_each(CL = c(0.5, 1, 1.5), KA = c(0.8, 1.1, 2))

}
