% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsolvetk.R
\name{sens_grid}
\alias{sens_grid}
\title{Sensitivity analysis with all parameter combinations.}
\usage{
sens_grid(mod, n = 100, ...)
}
\arguments{
\item{mod}{the model object}

\item{n}{the number of replicates to simulate}

\item{...}{named sequences of parameters; also arguments
passed to mrgsim}
}
\description{
Sensitivity analysis with all parameter combinations.
}
