% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsolvetk.R
\name{sens_norm}
\alias{sens_norm}
\title{Sensitivity analysis for log-normal parameters.}
\usage{
sens_norm(mod, n = 100, ...)
}
\arguments{
\item{mod}{the model object}
}
\description{
Sensitivity analysis for log-normal parameters.
}
