% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens_norm.R
\name{sens_norm_idata}
\alias{sens_norm_idata}
\title{Generate idata set for sens_norm}
\usage{
sens_norm_idata(pars, cv, .n = 100, ...)
}
\arguments{
\item{pars}{named numeric vector of parameters}

\item{cv}{coefficient of variation}

\item{.n}{number of replicates to simulate}

\item{...}{not used}
}
\description{
Generate idata set for sens_norm
}
\seealso{
\code{\link{sens_norm}}
}
