% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens_plot.R
\name{sens_plot}
\alias{sens_plot}
\title{Plot sensitivity analysis results}
\usage{
sens_plot(
  data,
  y,
  x = "time",
  col = split,
  log = FALSE,
  split = FALSE,
  lwd = 0.75
)
}
\arguments{
\item{data}{data frame of sensitivity analysis results}

\item{y}{name of the output variable to plot}

\item{x}{name of independent variable}

\item{col}{logical; if \code{TRUE} lines are colored by parameter value}

\item{log}{logical; if \code{TRUE}, y-axis is log-transformed}

\item{split}{logical; if \code{TRUE}, the data set is split on the}

\item{lwd}{passed to \code{geom_line} 
parameter name and a list of plots are returned; a single faceted plot is 
is returned otherwise}
}
\description{
Plot sensitivity analysis results
}
