% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens_range.R
\name{sens_range}
\alias{sens_range}
\alias{sens_range_factor}
\title{Sensitivity analysis at evenly spaced values}
\usage{
sens_range(mod, ..., .n = 5, .factor = NULL)

sens_range_factor(mod, ..., pars = names(param(mod)), .n = 5, .factor = 2)
}
\arguments{
\item{mod}{the model object}

\item{...}{named numeric vectors of length 2, specifying the minimum
and maximum value for that parameter}

\item{.n}{number of evenly spaced parameter values to simulate}

\item{.factor}{used to create a range for sensitivity analysis
based on the value of the selected parameters; the upper end 
of the range is \code{.factor} times the parameter value and the 
lower end of the range is the parameter value divided 
by \code{.factor}}

\item{pars}{a character vector of parameter names on which 
to do sensitivity analysis}
}
\description{
This is always a univariate (one parameter at a time) 
sensitivity analysis.
}
\examples{
mod <- mrgsolve:::house() \%>\% ev(amt = 100)

mod \%>\%
  sens_range(CL = c(0.5,1.5), VC = c(20,30), .n = 5)

}
