% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsolvetk.R
\name{sens_seq}
\alias{sens_seq}
\alias{sens_seq_}
\title{Sensitivity analysis with parameter sequences.}
\usage{
sens_seq(mod, n = 100, ...)

sens_seq_(mod, .dots)
}
\arguments{
\item{mod}{the model object}

\item{n}{the number of replicates to simulate}

\item{...}{named sequences of parameters; also arguments
passed to mrgsim}

\item{.dots}{list of arguments to pass to \code{\link{sens_seq}}}
}
\description{
Sensitivity analysis with parameter sequences.
}
\details{
In contrast to other simulation functions, 
\code{\link{sens_seq}} always returns 
data in long format with respect to 
the parameters involved in sensitivity
analysis.
}
\seealso{
\code{\link{sens_grid}}
}
