% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens_spaced.R
\name{sens_spaced}
\alias{sens_spaced}
\alias{sens_spaced_factor}
\title{Sensitivity analysis at evenly spaced values}
\usage{
sens_spaced(mod, ..., .n = 5, .factor = NULL)

sens_spaced_factor(mod, ..., pars = names(param(mod)), .n = 5,
  .factor = 2)
}
\arguments{
\item{mod}{the model object}

\item{...}{named numeric vectors of length 2, specifying the minimum
and maximum value for that parameter}

\item{.n}{number of evenly spaced parameter values to simulate}
}
\description{
This is always a univariate (one parameter at a time) 
sensitivity analysis.
}
\examples{
mod <- mrgsolve:::house() \%>\% ev(amt = 100)

mod \%>\%
  sens_spaced(CL = c(0.5,1.5), VC = c(20,30), .n = 5)

}
