% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsolvetk.R
\name{sens_unif}
\alias{sens_unif}
\alias{sens_unif_}
\title{Sensitivity analysis with uniform distribtion.}
\usage{
sens_unif(mod, n = 100, pars = names(param(mod)), ...)

sens_unif_(mod, .dots)
}
\arguments{
\item{mod}{the model object}

\item{n}{the number of replicates to simulate}

\item{pars}{character vector or comma-separated string of 
model parameters to simulate}

\item{...}{passed to \code{\link{sens_norm_idata}} and to mrgism}

\item{.dots}{list of arguments to pass to \code{\link{sens_unif}}}
}
\description{
Sensitivity analysis with uniform distribtion.
}
\details{
See the \code{spread} argument to \code{\link{sens_unif_idata}}.
}
\examples{
mod <- mrgsolve:::house()

out <- sens_unif(mod, n=10, pars="CL,VC")

}
\seealso{
\code{\link{sens_unif_idata}} \code{\link{covset}}
}
