% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsolvetk.R
\name{sens_unif_idata}
\alias{sens_unif_idata}
\title{Generate idata sets for sens_unif.}
\usage{
sens_unif_idata(pars, lower = 0.2, upper = 3, n = 100, spread = TRUE)
}
\arguments{
\item{pars}{character vector of parameter names}

\item{lower}{multiplier for lower bound}

\item{upper}{multiplier for upper bound}

\item{n}{number of replicates to simulate}

\item{spread}{if \code{TRUE} the data frame is returned in wide format}
}
\description{
Generate idata sets for sens_unif.
}
