% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsolvetk.R
\name{sim_lognorm}
\alias{sim_lognorm}
\title{Sensitivity analysis for log-normal parameters.}
\usage{
sim_lognorm(mod, niter = 100, ...)
}
\arguments{
\item{mod}{the model object}
}
\description{
Sensitivity analysis for log-normal parameters.
}
