% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgoptim.R
\name{mrgoptim}
\alias{mrgoptim}
\title{Fit a model to data using maximum likelihood objective function
with newuoa or optim}
\usage{
mrgoptim(
  mod,
  output = "ipred",
  var = "var",
  prms,
  v_prms,
  cov = TRUE,
  restarts = 0,
  method = c("newuoa", "bobyqa", "Nelder-Mead", "BFGS", "CG", "L-BFGS-B", "SANN",
    "Brent"),
  upper = Inf,
  lower = -Inf,
  options = switch(method, newuoa = list(iprint = 5), bobyqa = list(iprint = 5),
    list(trace = 2)),
  ...
)
}
\arguments{
\item{mod}{a model object}

\item{output}{a character string containing name of prediction output column}

\item{var}{a character string containing name of variance output column of predictions}

\item{prms}{a character vector of parameters to fit}

\item{v_prms}{a character vector of variance parameters to fit}

\item{cov}{logical, perform covariance step}

\item{restarts}{number of times to restart fit with final estimates of previous optimization}

\item{method}{method to be used: "newuoa", "bobyqa", "Nelder-Mead", "BFGS", "CG", "L-BFGS-B", "SANN", "Brent"}

\item{upper}{A numeric vector of upper bounds on the parameters. If the length is 1 the single upper bound is applied to all parameters. (only used for methods bobyqa, L-BFGS-B, Brent)}

\item{lower}{A numeric vector of lower bounds on the parameters. If the length is 1 the single lower bound is applied to all parameters. (only used for methods bobyqa, L-BFGS-B, Brent)}

\item{options}{options passed to optim function, see ?optim, ?newuoa, or ?bobyqa for more information}

\item{...}{further arguments passed to mrgsim function}
}
\description{
Fit a model to data using maximum likelihood objective function
with newuoa or optim
}
