% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgoptim.R
\name{mrgoptim}
\alias{mrgoptim}
\title{Fit a model to data using maximum likelihood objective function
with newuoa or optim}
\usage{
mrgoptim(mod, input, output, var, prms, v_prms, cov = TRUE, restarts = 0,
  method = c("newuoa", "Nelder-Mead", "BFGS", "CG", "L-BFGS-B", "SANN",
  "Brent"), options = list(maxit = 10000))
}
\arguments{
\item{mod}{a model object}

\item{input}{a numeric vector containing observation compartments in data_set to be fit}

\item{output}{a character vector containing names of pred output columns}

\item{var}{a character vector containing names of variance output columns of preds}

\item{prms}{a character vector of parameters to fit}

\item{v_prms}{a character vector of variance parameters to fit}

\item{cov}{logical, perform covariance step}

\item{restarts}{number of times to restart fit with final estimates of previous optimization}

\item{method}{method to be used: "newuoa","Nelder-Mead", "BFGS", "CG", "L-BFGS-B", "SANN", "Brent"}

\item{options}{options passed to optim function, see ?optim for more information}
}
\description{
Fit a model to data using maximum likelihood objective function
with newuoa or optim
}
