% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens_covset.R
\name{sens_covset}
\alias{sens_covset}
\title{Sensitivity analysis using covset objects}
\usage{
sens_covset(mod, covset, .n = 100, univariate = FALSE, ...)
}
\arguments{
\item{mod}{the model object}

\item{covset}{a covset object}

\item{.n}{the number of replicates to simulate}

\item{...}{passed to mutate_random and mrgsim}
}
\description{
Sensitivity analysis using covset objects
}
\examples{
mod <- mrgsolve:::house() \%>\% ev(amt = 100)

\dontrun{
library(dmutate)
cov1 <- covset(CL ~ runif(0.5, 1.5), VC ~ rlnorm(log(20), 0.5))
  
mod \%>\%
  sens_covset(.n = 10)
}

}
\seealso{
\code{\link{sens_unif}} \code{\link{sens_norm}}
}
