% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens_grid.R
\name{sens_grid}
\alias{sens_grid}
\title{Sensitivity analysis with all parameter combinations}
\usage{
sens_grid(mod, ...)
}
\arguments{
\item{mod}{the model object}

\item{...}{named sequences of parameters; also arguments
passed to mrgsim}
}
\description{
Sensitivity analysis with all parameter combinations
}
\details{
In contrast to other sensitivity analysis functions, 
\code{\link{sens_grid}} always returns data in 
wide format with respect to parameters involved
in the sensitivity analysis.
}
\examples{
mod <- mrgsolve:::house()
out <- 
  mod \%>\%
  ev(amt=100) \%>\%
  Req(CP) \%>\%
  sens_grid(CL = seq(1,2,0.2), VC = seq(10,40,5))
  
out


}
\seealso{
\code{\link{sens_seq}}
}
