% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsolvetk.R
\name{sens_norm}
\alias{sens_norm}
\alias{sens_norm_}
\title{Sensitivity analysis with log-normal distributions.}
\usage{
sens_norm(mod, n = 100, pars = names(param(mod)), ...)

sens_norm_(mod, .dots)
}
\arguments{
\item{mod}{the model object}

\item{n}{the number of replicates to simulate}

\item{pars}{character vector or comma-separated string of model
parameters to simulate}

\item{...}{passed to \code{\link{sens_norm_idata}} and to mrgism}

\item{.dots}{list of arguments to pass to \code{\link{sens_norm}}}
}
\description{
Sensitivity analysis with log-normal distributions.
}
