% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens_plot.R
\name{sens_plot}
\alias{sens_plot}
\title{Plot sensitivity analysis results}
\usage{
sens_plot(data, y, x = "time", col = split, log = FALSE, split = FALSE,
  lwd = 0.75)
}
\arguments{
\item{data}{data frame of sensitivity analysis results}

\item{col}{logical; if \code{TRUE} lines are colored by parameter value}

\item{log}{logical; if \code{TRUE}, y-axis is log-transformed}

\item{split}{logical; if \code{TRUE}, the data set is split on the 
parameter name and a list of plots are returned; a single faceted plot is 
is returned otherwise}

\item{ycol}{name of the output variable to plot}

\item{xcol}{name of independent variable}
}
\description{
Plot sensitivity analysis results
}
