% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens_seq.R
\name{sens_seq}
\alias{sens_seq}
\alias{sens_seq_}
\title{Sensitivity analysis with parameter sequences}
\usage{
sens_seq(mod, ...)

sens_seq_(mod, .dots)
}
\arguments{
\item{mod}{the model object}

\item{...}{named sequences of parameters; also arguments
passed to mrgsim}

\item{.dots}{list of arguments to pass to \code{\link{sens_seq}}}
}
\description{
Sensitivity analysis with parameter sequences
}
\details{
In contrast to other simulation functions, 
\code{\link{sens_seq}} always returns 
data in long format with respect to 
the parameters involved in sensitivity
analysis.
}
\examples{
mod <- mrgsolve:::house() \%>\% ev(amt = 100)

out <- 
  mod \%>\%
  sens_seq(CL = seq(1,2,0.2), VC = c(10,20,30))
  
out

}
\seealso{
\code{\link{sens_grid}}
}
