% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens_unif.R
\name{sens_unif_idata}
\alias{sens_unif_idata}
\title{Generate idata sets for sens_unif}
\usage{
sens_unif_idata(pars, lower = 1/upper, upper = 3, .n = 100, ...)
}
\arguments{
\item{pars}{named numeric vector of parameters}

\item{lower}{multiplier for lower bound}

\item{upper}{multiplier for upper bound}

\item{.n}{number of replicates to simulate}

\item{...}{not used}
}
\description{
Generate idata sets for sens_unif
}
\details{
It is important to note that \code{lower} and \code{upper} do
not correspond to the \code{min} and \code{max} arguments
for \code{\link{runif}}. Rather, they modify the current value
of the parameter in a multiplicative way.  For example, to 
simulate from uniform distributions that range from 
half the parameter value to double the parameter value,
use \code{lower} equal to \code{0.5} and \code{upper}
equal to \code{2}.
}
\examples{
pars <- c(CL = 1, VC = 2.2)

sens_unif_idata(pars, lower=0.67,upper=0.99, .n=5)


}
\seealso{
\code{\link{sens_unif}}
}
